Revert "config: change NTR constants to int"
This reverts commit d0a0e3aecd51b2829b51642f40b7b03bd6fcc8f9.

+++ config/Config.hpp
@@ -14,10 +14,10 @@ public:
 
     struct NetworkConfig {
         std::string host = "";
-        uint priority = 1;
-        uint priorityFactor = 5;
-        uint jpegQuality = 80;
-        uint QoS = 18;
+        uint8_t priority = 1;
+        uint8_t priorityFactor = 5;
+        uint8_t jpegQuality = 80;
+        uint8_t QoS = 18;
 
         int input_ratelimit_us = 50 * 1000;
         int input_pollrate_us = 5 * 1000;
